+++ front/scripts/main/components/SiteHeadComponent.ts
@@ -25,10 +25,11 @@ App.SiteHeadComponent = Em.Component.extend(App.TrackClickMixin, {
 
 	actions: {
 		expandSideNav: function (): void {
-			this.set('sideNavCollapsed', false);
+			this.sendAction('toggleSideNav', true);
 		},
+
 		showUserMenu: function (): void {
-			this.set('userMenuCollapsed', false);
+			this.sendAction('toggleUserMenu', true);
 		}
 	},
 
@@ -39,8 +40,10 @@ App.SiteHeadComponent = Em.Component.extend(App.TrackClickMixin, {
 	smartBannerVisibleObserver: Em.observer('smartBannerVisible', function (): void {
 		var headroom = this.get('headroom');
 
-		headroom.destroy();
-		this.initHeadroom();
+		if (headroom) {
+			headroom.destroy();
+			this.initHeadroom();
+		}
 	}),
 
 	/**