+++ front/scripts/main/controllers/ArticleController.ts
@@ -16,11 +16,6 @@ App.ArticleController = Em.Controller.extend({
 	},
 
 	actions: {
-		updateHeaders: function (headers: NodeList): void {
-			var article = this.get('model');
-			article.set('sections', headers);
-		},
-
 		edit: function (title: string, sectionIndex: number): void {
 			App.VisibilityStateManager.reset();
 			this.transitionToRoute('edit', title, sectionIndex);
@@ -32,8 +27,17 @@ App.ArticleController = Em.Controller.extend({
 			});
 		},
 
-		addPhoto: function(title: string, sectionIndex: number, photoData: any): void {
-			this.transitionToRoute('addPhoto', App.AddPhotoModel.load(title, sectionIndex, photoData));
+		addPhoto: function (title: string, sectionIndex: number, photoData: any): void {
+			var photoModel = App.AddPhotoModel.load(photoData);
+			//We don't want to hold with transition and wait for a promise to resolve.
+			//Instead we set properties on model after resolving promise and Ember scheduler handles this gracefully.
+			photoModel.then((model: typeof App.AddPhotoModel) => {
+				model.setProperties({
+					title,
+					sectionIndex
+				});
+			});
+			this.transitionToRoute('addPhoto', photoModel);
 		},
 
 		articleRendered: function (): void {