+++ front/scripts/main/routes/CuratedContentEditorSectionEditRoute.ts
@@ -1,0 +1,45 @@
+/// <reference path="../app.ts" />
+/// <reference path="../../../../typings/ember/ember.d.ts" />
+
+'use strict';
+
+App.CuratedContentEditorSectionEditRoute = Em.Route.extend({
+	model(): CuratedContentEditorItemModel {
+		return $.extend(true, {}, this.modelFor('curatedContentEditor.section'));
+	},
+
+	setupController(controller: any, model: CuratedContentEditorItemModel, transition: EmberStates.Transition): void {
+		var parentSection = this.modelFor('curatedContentEditor').get('curated');
+
+		this._super(controller, model, transition);
+		controller.set('alreadyUsedLabels', App.CuratedContentEditorModel.getAlreadyUsedLabels(parentSection, model.label));
+	},
+
+	renderTemplate(): void {
+		this.render('curated-content-editor-item');
+	},
+
+	actions: {
+		goBack(): void {
+			var sectionModel: CuratedContentEditorItemModel = this.modelFor('curatedContentEditor.section');
+			this.transitionTo('curatedContentEditor.section', encodeURIComponent(sectionModel.label));
+		},
+
+		done(newSection: CuratedContentEditorItemModel): void {
+			var sectionModel: CuratedContentEditorItemModel = this.modelFor('curatedContentEditor.section');
+
+			Em.setProperties(sectionModel, newSection);
+			this.transitionTo('curatedContentEditor.section.index');
+		},
+
+		// Delete section
+		deleteItem(): void {
+			var rootModel: CuratedContentEditorModel = this.modelFor('curatedContentEditor'),
+				controller: any = this.controllerFor('curatedContentEditor.section'),
+				originalSectionLabel: string = controller.get('originalSectionLabel');
+
+			App.CuratedContentEditorModel.deleteItem(rootModel['curated'], originalSectionLabel);
+			this.transitionTo('curatedContentEditor.index');
+		}
+	}
+});