+++ homepage/server/routes.js
@@ -1,0 +1,50 @@
+/**
+ * Wikia (Japan) Homepage
+ *
+ * @author Per Johan Groland <pgroland@wikia-inc.com>
+ */
+
+'use strict';
+
+exports.routes = [
+	{
+		method: 'GET',
+		path: '/',
+		handler: require('./facets/index')
+	},
+	{
+		method: 'GET',
+		path: '/beginners',
+		handler: require('./facets/beginners')
+	},
+	{
+		method: 'GET',
+		path: '/companyinfo',
+		handler: require('./facets/companyinfo')
+	},
+	{
+		method: 'GET',
+		path: '/contact',
+		handler: require('./facets/contact')
+	},
+	{
+		method: 'GET',
+		path: '/login',
+		handler: require('./facets/signin')
+	},
+	{
+		method: 'GET',
+		path: '/privacy',
+		handler: require('./facets/privacy')
+	},
+	{
+		method: 'GET',
+		path: '/termsofservice',
+		handler: require('./facets/termsofservice')
+	},
+	{
+		method: 'GET',
+		path: '/{path*}',
+		handler: require('./facets/assets')
+	}
+];